requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fiona==1.9.6
fsspec==2024.10.0
geographiclib==2.0
geopandas==0.13.2
geopy==2.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.52.2
packaging==24.1
pandas==2.0.3
pydantic==2.9.2
pydantic_core==2.23.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.20.2
